requirements.txt
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
ddgs>=9.10.0
anthropic>=0.83.0
chromadb>=1.0.0
